♻️ Move lnschema-core modules into base submodule
