+++ lamindb/_artifact.py
@@ -11,6 +11,9 @@ import lamindb_setup as ln_setup
 import pandas as pd
 from anndata import AnnData
 from django.db.models import Q
+from lamidb.base.types import (
+    VisibilityChoice,
+)
 from lamin_utils import colors, logger
 from lamindb_setup import settings as setup_settings
 from lamindb_setup._init_instance import register_storage_in_instance
@@ -25,9 +28,6 @@ from lamindb_setup.core.upath import (
 )
 
 from lamindb.models import Artifact, FeatureManager, ParamManager, Run, Storage
-from lamindb.types import (
-    VisibilityChoice,
-)
 
 from ._parents import view_lineage
 from ._utils import attach_func_to_class_method
