+++ lamindb/_feature_set.py
@@ -4,13 +4,13 @@ from typing import TYPE_CHECKING
 
 import lamindb_setup as ln_setup
 import numpy as np
+from lamidb.base.types import FieldAttr, ListLike
 from lamin_utils import logger
 from lamindb_setup.core._docs import doc_args
 from lamindb_setup.core.hashing import hash_set
 
-from lamindb import ids
+from lamindb.base import ids
 from lamindb.models import Feature, FeatureSet, Record
-from lamindb.types import FieldAttr, ListLike
 
 from ._feature import convert_pandas_dtype_to_lamin_dtype
 from ._record import init_self_from_db
