+++ lamindb/_parents.py
@@ -20,8 +20,9 @@ from ._record import get_name_field
 from ._utils import attach_func_to_class_method
 
 if TYPE_CHECKING:
+    from lamidb.base.types import StrField
+
     from lamindb.core import QuerySet
-    from lamindb.types import StrField
 
 LAMIN_GREEN_LIGHTER = "#10b981"
 LAMIN_GREEN_DARKER = "#065f46"
