+++ lamindb/_query_manager.py
@@ -12,7 +12,7 @@ from .core._feature_manager import get_feature_set_by_slot_
 from .core._settings import settings
 
 if TYPE_CHECKING:
-    from lamindb.types import StrField
+    from lamidb.base.types import StrField
 
 
 class QueryManager(models.Manager):
