+++ lamindb/models.py
@@ -22,12 +22,21 @@ from django.db.models.fields.related import (
     ManyToManyRel,
     ManyToOneRel,
 )
+from lamidb.base.types import (
+    ArtifactType,
+    FeatureDtype,
+    FieldAttr,
+    ListLike,
+    StrField,
+    TransformType,
+    VisibilityChoice,
+)
 from lamin_utils import colors
 from lamindb_setup import _check_instance_setup
 from lamindb_setup.core._docs import doc_args
 from lamindb_setup.core.hashing import HASH_LENGTH
 
-from lamindb.fields import (
+from lamindb.base.fields import (
     BigIntegerField,
     BooleanField,
     CharField,
@@ -36,15 +45,6 @@ from lamindb.fields import (
     IntegerField,
     OneToOneField,
     TextField,
-)
-from lamindb.types import (
-    ArtifactType,
-    FeatureDtype,
-    FieldAttr,
-    ListLike,
-    StrField,
-    TransformType,
-    VisibilityChoice,
 )
 
 from .ids import base62_8, base62_12, base62_20
@@ -57,6 +57,13 @@ if TYPE_CHECKING:
     import numpy as np
     import pandas as pd
     from anndata import AnnData
+    from lamidb.base.mocks import (
+        AnnDataAccessor,
+        BackedAccessor,
+        MappedCollection,
+        QuerySet,
+        RecordList,
+    )
     from lamin_utils._inspect import InspectResult
     from lamindb_setup.core.types import UPathStr
     from mudata import MuData
@@ -66,13 +73,6 @@ if TYPE_CHECKING:
     from upath import UPath
 
     from lamindb.core import LabelManager
-    from lamindb.mocks import (
-        AnnDataAccessor,
-        BackedAccessor,
-        MappedCollection,
-        QuerySet,
-        RecordList,
-    )
 
 
 _TRACKING_READY: bool | None = None
